goodreads/pyproject.toml
[tool.poetry]
name = "goodreads"
version = "0.1.0"
description = ""
authors = ["Kavie <moklavie@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
djangorestframework-jsonapi = "^4.0.0"
python-environ = "^0.4.54"
psycopg2-binary = "^2.8.6"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
